Modificando calculo de distancia

+++ Utils/GeoUtils.cs
@@ -7,14 +7,14 @@ namespace Prueba_Geolocalizacion.Utils
 {
     public class GeoUtils
     {
-        public static double CalcularDistancia(double lat1, double lon1, double lat2, double lon2)
+        public static double CalcularDistancia(decimal lat1, decimal lon1, decimal lat2, decimal lon2)
         {
             const double R = 6371e3; // radio de la Tierra en metros
 
-            var φ1 = DegreesToRadians(lat1);
-            var φ2 = DegreesToRadians(lat2);
-            var Δφ = DegreesToRadians(lat2 - lat1);
-            var Δλ = DegreesToRadians(lon2 - lon1);
+            var φ1 = DegreesToRadians((double)lat1);
+            var φ2 = DegreesToRadians((double)lat2);
+            var Δφ = DegreesToRadians((double)(lat2 - lat1));
+            var Δλ = DegreesToRadians((double)(lon2 - lon1));
 
             var a = Math.Sin(Δφ / 2) * Math.Sin(Δφ / 2) +
                     Math.Cos(φ1) * Math.Cos(φ2) *
